refactor(core): don't dynamically create element classes
This simplifies using the paging, an Annotation element is always an
Annotation element, otherwise two elements from two different instances
of a namespace differed
