+++ packages/api-elements/lib/elements/Annotation.js
@@ -1,0 +1,27 @@
+const { StringElement } = require('minim');
+
+/**
+ * @class Annotation
+ *
+ * @param {string} content
+ * @param meta
+ * @param attributes
+ *
+ * @extends StringElement
+ */
+class Annotation extends StringElement {
+  constructor(...args) {
+    super(...args);
+    this.element = 'annotation';
+  }
+
+  get code() {
+    return this.attributes.get('code');
+  }
+
+  set code(value) {
+    this.attributes.set('code', value);
+  }
+}
+
+module.exports = Annotation;
