--- packages/api-elements/lib/elements/copy.js
@@ -1,48 +1,0 @@
-module.exports = (namespace) => {
-  const StringElement = namespace.getElementClass('string');
-  const ArrayElement = namespace.getElementClass('array');
-
-  /**
-   * @name copy
-   * @type Copy
-   * @memberof Element.prototype
-   */
-  if (!Object.getOwnPropertyNames(ArrayElement.prototype).includes('copy')) {
-    Object.defineProperty(ArrayElement.prototype, 'copy', {
-      get() {
-        return this.children.filter(item => item.element === 'copy');
-      },
-    });
-  }
-
-  /**
-   * @class Copy
-   *
-   * @param {string} content
-   * @param meta
-   * @param attributes
-   *
-   * @extends StringElement
-   */
-  class Copy extends StringElement {
-    constructor(...args) {
-      super(...args);
-      this.element = 'copy';
-    }
-
-    /**
-     * @name contentType
-     * @type StringElement
-     * @memberof Copy.prototype
-     */
-    get contentType() {
-      return this.attributes.get('contentType');
-    }
-
-    set contentType(value) {
-      this.attributes.set('contentType', value);
-    }
-  }
-
-  namespace.register('copy', Copy);
-};
