--- packages/api-elements/lib/elements/href-variables.js
@@ -1,21 +1,0 @@
-module.exports = (namespace) => {
-  const ObjectElement = namespace.getElementClass('object');
-
-  /**
-   * @class HrefVariables
-   *
-   * @param {Array} content
-   * @param meta
-   * @param attributes
-   *
-   * @extends ObjectElement
-   */
-  class HrefVariables extends ObjectElement {
-    constructor(...args) {
-      super(...args);
-      this.element = 'hrefVariables';
-    }
-  }
-
-  namespace.register('hrefVariables', HrefVariables);
-};
