--- packages/api-elements/lib/elements/http-transaction.js
@@ -1,48 +1,0 @@
-module.exports = (namespace) => {
-  const ArrayElement = namespace.getElementClass('array');
-
-  /**
-   * @class HttpTransaction
-   *
-   * @param {Array} content
-   * @param meta
-   * @param attributes
-   *
-   * @extends ArrayElement
-   */
-  class HttpTransaction extends ArrayElement {
-    constructor(...args) {
-      super(...args);
-      this.element = 'httpTransaction';
-    }
-
-    /**
-     * @name request
-     * @type HttpRequest
-     * @memberof HttpTransaction.prototype
-     */
-    get request() {
-      return this.children.filter(item => item.element === 'httpRequest').first;
-    }
-
-    /**
-     * @name response
-     * @type HttpResponse
-     * @memberof HttpTransaction.prototype
-     */
-    get response() {
-      return this.children.filter(item => item.element === 'httpResponse').first;
-    }
-
-    /**
-     * @name authSchemes
-     * @type ArrayElement
-     * @memberof HttpTransaction.prototype
-     */
-    get authSchemes() {
-      return this.attributes.get('authSchemes');
-    }
-  }
-
-  namespace.register('httpTransaction', HttpTransaction);
-};
