+++ packages/api-elements/lib/parse-result.js
@@ -8,15 +8,17 @@
  */
 
 const apiDescription = require('./api-description');
-const parseResult = require('./elements/parse-result');
-const annotation = require('./elements/annotation');
-const sourceMap = require('./elements/source-map');
+const ParseResult = require('./elements/ParseResult');
+const Annotation = require('./elements/Annotation');
+const SourceMap = require('./elements/SourceMap');
 const defineSourceMapValue = require('./define-source-map-value');
 
 const namespace = (options) => {
-  parseResult(options.base);
-  annotation(options.base);
-  sourceMap(options.base);
+  options.base
+    .register('parseResult', ParseResult)
+    .register('annotation', Annotation)
+    .register('sourceMap', SourceMap);
+
   defineSourceMapValue(options.base);
 
   options.base.use(apiDescription);
